requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-dotenv
pypdf
openai
